pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Dmitriy_Bessalov <dmitriy@bessalov.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = {extras = ["standard"], version = "^0.17.6"}
PyYAML = "^6.0"
SQLAlchemy = "^1.4.34"
databases = "^0.5.5"
asyncpg = "^0.25.0"
alembic = "^1.7.7"
psycopg2 = "^2.9.3"
sqlmodel = "^0.0.6"
fastapi = "^0.78.0"
strawberry-graphql = "^0.113.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
